third_party/WebKit/LayoutTests/fast/dom/node-legacy-event-listener-expected.txt

Test addEventListener() and removeEventListener() fail silently if arguments are missing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.addEventListener("foo") threw exception TypeError: Failed to execute 'addEventListener' on 'EventTarget': 2 arguments required, but only 1 present..
PASS document.removeEventListener("bar") threw exception TypeError: Failed to execute 'removeEventListener' on 'EventTarget': 2 arguments required, but only 1 present..
PASS document.addEventListener() threw exception TypeError: Failed to execute 'addEventListener' on 'EventTarget': 2 arguments required, but only 0 present..
PASS document.removeEventListener() threw exception TypeError: Failed to execute 'removeEventListener' on 'EventTarget': 2 arguments required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
